# feat(performance): Optimize ControlPanel2 and Bomber Tab