--- a/src/client/events/PlayerListChangedEvent.ts
+++ b/src/client/events/PlayerListChangedEvent.ts
@@ -1,0 +1,6 @@
+import { GameEvent } from "../../core/EventBus";
+
+/**
+ * Fired when the list of living players changes (join, leave, or defeat).
+ */
+export class PlayerListChangedEvent implements GameEvent {}