## Description:

#### **Summary**

This pull request addresses significant performance bottlenecks in the `ControlPanel2` component, dramatically improving UI responsiveness, especially in large games. The primary issues stemmed from inefficient, polling-based calculations in the "Bombers" tab and for player list updates, which caused high CPU load on every game tick.

#### **Changes Implemented**

This PR introduces three key optimizations:

1.  **Spatial Index for Bomber Targeting:**
    *   The `_getPlayersInAirfieldRange()` method, which calculates targets for bombers, has been completely refactored.
    *   It now uses a `SpatialIndex` for structure lookups, replacing the previous brute-force iteration over all players and structures. This reduces the complexity of range calculations and provides a massive performance boost.

2.  **Event-Driven Player List Updates:**
    *   The component no longer polls for changes to the game's player list on every tick.
    *   It now subscribes to a new `PlayerListChangedEvent`, ensuring that UI updates related to players joining or leaving only happen when necessary, further reducing the per-tick workload.

3.  **Regression Fix for Unnecessary Calculations:**
    *   A regression was identified where the bomber range calculation was running on every tick, even when the "Bombers" tab was closed.
    *   This has been corrected by wrapping the logic in a conditional check, ensuring the expensive calculation only runs when the relevant UI is visible.

---

### **Performance Improvements**

Manual in-game testing was conducted to measure the impact of these changes. The results show a dramatic improvement in performance.

#### **Bomber Tab OPEN (Baseline vs. Optimized)**

The optimization is a **major success**, with the average `tick()` time being approximately **7 times faster**.

| Metric                               |  Baseline | Optimized |                      Change |
| ------------------------------------ | --------: | --------: | --------------------------: |
| `tick()` avg                         | 0.8075 ms |  0.116 ms |    **85.6% (\~7 x faster)** |
| `_getPlayersInAirfieldRange()` avg   | 0.6665 ms |  0.058 ms | **91.3% (\~11.5 x faster)** |

<br>

#### **Bomber Tab CLOSED (Regression & Fix)**

The fix for the regression was highly effective, **eliminating all unnecessary calls** and making the `tick()` **twice as fast** as it was before the fix.

| Metric                               | Before Fix | After Fix | Change |
| ------------------------------------ | ---------: | --------: | --------------------------: |
| `tick()` avg                         |   0.103 ms | **~0.052 ms** |                  **49.5%** |
| `_getPlayersInAirfieldRange()` calls |       1000 |       **0** |                  **100%** |

<br>

These changes make the UI significantly more performant and scalable for larger games.

## Please complete the following:

- [x] I have added screenshots for all UI updates
- [x] I process any text displayed to the user through translateText() and I've added it to the en.json file
- [x] I have added relevant tests to the test directory
- [x] I confirm I have thoroughly tested these changes and take full responsibility for any bugs introduced
- [x] I understand that submitting code with bugs that could have been caught through manual testing blocks releases and new features for all contributors

## Please put your Discord username so you can be contacted if a bug or regression is found:

el_magico777